Simplify TelemetrySourceAdapater part 1
- Get rid of abstract class
- Get rid of service registration
