+++ src/Microsoft.Framework.TelemetryAdapter/TelemetryListenerExtensions.cs
@@ -10,9 +10,9 @@ namespace System.Diagnostics.Tracing
         public static IDisposable SubscribeWithAdapter(this TelemetryListener telemetry, object target)
         {
 #if PROXY_SUPPORT
-            var adapter = new DefaultTelemetrySourceAdapter(new ProxyTelemetrySourceMethodAdapter());
+            var adapter = new TelemetrySourceAdapter(new ProxyTelemetrySourceMethodAdapter());
 #else
-            var adapter = new DefaultTelemetrySourceAdapter(new ReflectionTelemetrySourceMethodAdapter());
+            var adapter = new TelemetrySourceAdapter(new ReflectionTelemetrySourceMethodAdapter());
 #endif
             adapter.EnlistTarget(target);
             return telemetry.Subscribe(adapter, adapter.IsEnabled);
