+++ src/Microsoft.Framework.TelemetryAdapter/TelemetrySourceAdapter.cs
@@ -1,12 +1,132 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-using System.Diagnostics.Tracing;
+using System;
+using System.Collections.Concurrent;
+using System.Collections.Generic;
+using System.Diagnostics;
+using System.Reflection;
 
 namespace Microsoft.Framework.TelemetryAdapter
 {
-    public abstract class TelemetrySourceAdapter : TelemetrySource
+    public class TelemetrySourceAdapter : IObserver<KeyValuePair<string, object>>
     {
-        public abstract void EnlistTarget(object listener);
+        private readonly ListenerCache _listeners = new ListenerCache();
+        
+        private readonly ITelemetrySourceMethodAdapter _methodAdapter;
+
+        public TelemetrySourceAdapter(ITelemetrySourceMethodAdapter methodAdapter)
+        {
+            _methodAdapter = methodAdapter;
+        }
+
+        public void EnlistTarget(object target)
+        {
+            var typeInfo = target.GetType().GetTypeInfo();
+
+            var methodInfos = typeInfo.DeclaredMethods;
+
+            foreach (var methodInfo in methodInfos)
+            {
+                var notificationNameAttribute = methodInfo.GetCustomAttribute<TelemetryNameAttribute>();
+                if (notificationNameAttribute != null)
+                {
+                    Enlist(notificationNameAttribute.Name, target, methodInfo);
+                }
+            }
+        }
+
+        private void Enlist(string notificationName, object target, MethodInfo methodInfo)
+        {
+            var entries = _listeners.GetOrAdd(
+                notificationName,
+                _ => new ConcurrentBag<ListenerEntry>());
+
+            entries.Add(new ListenerEntry(target, methodInfo));
+        }
+
+        public bool IsEnabled(string telemetryName)
+        {
+            if (_listeners.Count == 0)
+            {
+                return false;
+            }
+
+            return _listeners.ContainsKey(telemetryName);
+        }
+
+        public void WriteTelemetry(string telemetryName, object parameters)
+        {
+            if (parameters == null)
+            {
+                return;
+            }
+
+            ConcurrentBag<ListenerEntry> entries;
+            if (_listeners.TryGetValue(telemetryName, out entries))
+            {
+                foreach (var entry in entries)
+                {
+                    var succeeded = false;
+                    foreach (var adapter in entry.Adapters)
+                    {
+                        if (adapter(entry.Target, parameters))
+                        {
+                            succeeded = true;
+                            break;
+                        }
+                    }
+
+                    if (!succeeded)
+                    {
+                        var newAdapter = _methodAdapter.Adapt(entry.MethodInfo, parameters.GetType());
+                        succeeded = newAdapter(entry.Target, parameters);
+                        Debug.Assert(succeeded);
+
+                        entry.Adapters.Add(newAdapter);
+                    }
+                }
+            }
+        }
+
+        void IObserver<KeyValuePair<string, object>>.OnNext(KeyValuePair<string, object> value)
+        {
+            WriteTelemetry(value.Key, value.Value);
+        }
+
+        void IObserver<KeyValuePair<string, object>>.OnError(Exception error)
+        {
+            // Do nothing
+        }
+
+        void IObserver<KeyValuePair<string, object>>.OnCompleted()
+        {
+            // Do nothing
+        }
+
+        private class ListenerCache : ConcurrentDictionary<string, ConcurrentBag<ListenerEntry>>
+        {
+            public ListenerCache()
+                : base(StringComparer.Ordinal)
+            {
+            }
+        }
+
+        private class ListenerEntry
+        {
+            public ListenerEntry(object target, MethodInfo methodInfo)
+            {
+                Target = target;
+                MethodInfo = methodInfo;
+
+                Adapters = new ConcurrentBag<Func<object, object, bool>>();
+            }
+
+            public ConcurrentBag<Func<object, object, bool>> Adapters { get; }
+
+            public MethodInfo MethodInfo { get; }
+
+            public object Target { get; }
+        }
     }
 }
